* devel-2.1:
  Convert gnt-instance list and info to use roman
  gnt-cluster info --roman
  FormatUidPool: provide optional roman conversion
  gnt-node: remove latinfriendlyfields
  Move roman conversion to compat
  Add a new opcode timestamp field
  Fix IgnoreSignals on socket.error
  RAPI client should convert urllib2.URLError to GanetiApiError
  KVM: Migration bandwidth and downtime control
  Make utils.EnsureDirs() ignore umask
  Fix two race conditions in reboot instance
  Support for latin friendly output in node list

Conflicts:
	man/gnt-instance.sgml
	  - trivial

+++ lib/hypervisor/hv_xen.py
@@ -213,6 +213,10 @@ class XenHypervisor(hv_base.BaseHypervisor):
     """
     ini_info = self.GetInstanceInfo(instance.name)
 
+    if ini_info is None:
+      raise errors.HypervisorError("Failed to reboot instance %s,"
+                                   " not running" % instance.name)
+
     result = utils.RunCmd(["xm", "reboot", instance.name])
     if result.failed:
       raise errors.HypervisorError("Failed to reboot instance %s: %s, %s" %
@@ -223,7 +227,8 @@ class XenHypervisor(hv_base.BaseHypervisor):
       new_info = self.GetInstanceInfo(instance.name)
 
       # check if the domain ID has changed or the run time has decreased
-      if new_info[1] != ini_info[1] or new_info[5] < ini_info[5]:
+      if (new_info is not None and
+          (new_info[1] != ini_info[1] or new_info[5] < ini_info[5])):
         return
 
       raise utils.RetryAgain()
